Support more than a single change happening in a second
If two or more changes are made withing a single second it is possible
that only the first change will be loaded due to the granularity of
st_mtime.  Work around this by zeroing the saved mtime when we get
a poke on the socket from crontab or at since that means there has been
a change.  Fix from Daniel Lucq.