Allow crontab files to be mode 0600, looser restrictions on /etc/crontab
Since it not edited via crontab(1) there's no good reason to require
its mode to be be 0400/0600 or to bail if the link count is != 1.
The reason for such paranoia all has to do with crontab(1) and
editing user crontab files.  We now only require that /etc/crontab
not be writable by anyone other than the owner.  From Walt Howard.