CVE-2010-0424: use futimens() if possible or run utime() as user
For systems without futimens(), call utime() as the user (not root)
to prevent an attack where the user replaces the temporary file
being edited with a symbolic link to another file.  Since utime()
was run as root, the user could change the time stamp on any file
to that of the crontab file.  Also move TEQUAL to macros.h so we
don't have to define it in multiple places.